+++ lib/ClassMover/Tests/Adapter/TolerantParser/examples/Example10.php
@@ -1,0 +1,10 @@
+<?php
+
+use Foobar\Example as BadExample;
+
+class ClassOne
+{
+    public function build(): BadExample
+    {
+    }
+}